[metrics]fix lastRate & meanRate long/double problem

+++ modules/metrics/src/main/java/org/springside/modules/metrics/CounterMetric.java
@@ -7,9 +7,9 @@ package org.springside.modules.metrics;
 
 public class CounterMetric {
 	public long totalCount;
-	public double meanRate;
+	public long meanRate;
 	public long lastCount;
-	public double lastRate;
+	public long lastRate;
 
 	@Override
 	public String toString() {
